% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_bootstrap.R
\name{bootstrap_parallel}
\alias{bootstrap_parallel}
\title{Runs bootstrap using a starting point estimate (parallel implementation)
(Algorithm 4 from paper)}
\usage{
bootstrap_parallel(
  data_cov,
  data_outcomes,
  theta_init,
  beta_init,
  M_mar,
  s_max,
  numBootstraps = 1000,
  stop_eps_mar = 1e-04,
  stop_eps_latent = 1e-04,
  num_imps = 20
)
}
\arguments{
\item{data_cov}{n x p matrix of covariates}

\item{data_outcomes}{n x d matrix of discrete outcomes}

\item{theta_init}{K x d matrix of binomial initial params}

\item{beta_init}{(K-1) x (p+1) matrix of beta (weight) initial params}

\item{M_mar}{number of iterations of the outer EM algorithm}

\item{s_max}{vector of length d, containing the maximum score for each test}

\item{numBootstraps}{number of bootstrap samples}

\item{stop_eps_mar}{epsilon for checking convergence for each outer EM iter}

\item{stop_eps_latent}{epsilon for checking convergence for each inner EM iter}

\item{num_imps}{number of multiple imputations}
}
\value{
A list containing the bootstrap estimates for beta and theta
}
\description{
`bootstrap_parallel()` returns the bootstrap estimates for the parameters
}
