% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{changeRef}
\alias{changeRef}
\title{Shifts the theta and beta parameters to account for changing class reference}
\usage{
changeRef(theta, beta)
}
\arguments{
\item{theta}{K x d matrix of binomial parameters}

\item{beta}{(K-1) x (p+1) matrix of beta (weight) parameters}
}
\value{
A list containing the updated theta and beta matrices
}
\description{
`changeRef()` returns new theta and beta parameters, shifted, so that the
new reference class is such that `theta[,1]' is maximized
}
