% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gendata.R
\name{expit}
\alias{expit}
\title{Computes expit of a vector}
\usage{
expit(input)
}
\arguments{
\item{input}{numerical vector}
}
\value{
A vector containing expit(input)
}
\description{
`expit()` returns expit of a vector
}
