% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_est.R
\name{fit_latent_var}
\alias{fit_latent_var}
\title{Fit the mixture of binomial product experts model when there is no missingness
(Algorithm 1 from paper)}
\usage{
fit_latent_var(
  data_cov,
  data_outcomes,
  theta_init,
  beta_init,
  K,
  M_em = 50,
  s_max,
  printFlag = FALSE,
  stop_eps = 1e-04
)
}
\arguments{
\item{data_cov}{n x p matrix of covariates}

\item{data_outcomes}{n x d matrix of discrete outcomes}

\item{theta_init}{K x d matrix of binomial initial params}

\item{beta_init}{(K-1) x (p+1) matrix of beta (weight) initial params}

\item{K}{number of mixture components}

\item{M_em}{number of iterations of the EM algorithm}

\item{s_max}{vector of length d, containing the maximum score for each test}

\item{printFlag}{True/False flag to print status}

\item{stop_eps}{epsilon for checking convergence}
}
\value{
A list containing the beta_star,
theta_star, and log_likelihood evaluated at the convergence point
}
\description{
`fit_latent_var()` returns a list containing the beta_star,
theta_star, and log_likelihood evaluated at the convergence point
}
