% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_est.R
\name{fit_mar}
\alias{fit_mar}
\title{Fit the mixture of binomial product experts model when there is missingness
under MAR assumption
(Algorithm 3 from paper)}
\usage{
fit_mar(
  data_cov,
  data_outcomes,
  theta_init,
  beta_init,
  K,
  M_mar = 50,
  s_max,
  printFlag = FALSE,
  stop_eps_mar = 1e-04,
  stop_eps_latent = 1e-04,
  num_imps = 20
)
}
\arguments{
\item{data_cov}{n x p matrix of covariates}

\item{data_outcomes}{n x d matrix of discrete outcomes}

\item{theta_init}{K x d matrix of binomial initial params}

\item{beta_init}{(K-1) x (p+1) matrix of beta (weight) initial params}

\item{K}{number of mixture components}

\item{M_mar}{number of iterations of the outer EM algorithm}

\item{s_max}{vector of length d, containing the maximum score for each test}

\item{printFlag}{True/False flag to print status}

\item{stop_eps_mar}{epsilon for checking convergence for each outer EM iter}

\item{stop_eps_latent}{epsilon for checking convergence for each inner EM iter}

\item{num_imps}{number of multiple imputations}
}
\value{
A list containing the beta_star,
theta_star, and log_likelihood evaluated at the convergence point
}
\description{
`fit_mar()` returns a list containing the beta_star,
theta_star, and log_likelihood evaluated at the convergence point
}
