% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_est.R
\name{fit_mar_randinit}
\alias{fit_mar_randinit}
\title{Fit the mixture of binomial product experts model when there is missingness
under MAR assumption with random initializations (serial implementation)
(Algorithm 3 from paper)}
\usage{
fit_mar_randinit(
  data_cov,
  data_outcomes,
  numRandInit,
  K,
  M_mar,
  s_max,
  printFlag = FALSE,
  stop_eps_mar = 1e-04,
  stop_eps_latent = 1e-04,
  num_imps = 20
)
}
\arguments{
\item{data_cov}{n x p matrix of covariates}

\item{data_outcomes}{n x d matrix of discrete outcomes}

\item{numRandInit}{number of random initializations}

\item{K}{number of mixture components}

\item{M_mar}{number of iterations of the outer EM algorithm}

\item{s_max}{vector of length d, containing the maximum score for each test}

\item{printFlag}{True/False flag to print status}

\item{stop_eps_mar}{epsilon for checking convergence for each outer EM iter}

\item{stop_eps_latent}{epsilon for checking convergence for each inner EM iter}

\item{num_imps}{number of multiple imputations}
}
\value{
A list containing the beta_star,
theta_star, and log_likelihood evaluated at the convergence point
}
\description{
`fit_mar_randinit()` returns a list containing the beta_star,
theta_star, and log_likelihood evaluated at the convergence point
}
