% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gendata.R
\name{generate_cc_data}
\alias{generate_cc_data}
\title{Generates a complete case data set without missing data}
\usage{
generate_cc_data(n)
}
\arguments{
\item{n}{number of observations}
}
\value{
A list of X (n x p) covariate matrix, Y (n x d) discrete outcome
matrix, Z (n x 1) vector of class labels
}
\description{
`generate_cc_data()` returns a generated data set
}
