% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{get_loglikelihood}
\alias{get_loglikelihood}
\title{Computes the observed log-likelihood}
\usage{
get_loglikelihood(data_cov, data_outcomes, s_max, theta, beta)
}
\arguments{
\item{data_cov}{n x p matrix of covariates}

\item{data_outcomes}{n x d matrix of discrete outcomes}

\item{s_max}{vector of length d, containing the maximum score for each test}

\item{theta}{K x d matrix of binomial parameters}

\item{beta}{(K-1) x (p+1) matrix of beta (weight) parameters}
}
\value{
The observed log-likelihood value
}
\description{
`get_loglikelihood()` computes the observed log-likelihood at (theta, beta)
using the data
}
