% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{get_R_matrix}
\alias{get_R_matrix}
\title{Obtain the missing pattern matrix R}
\usage{
get_R_matrix(data_outcomes, s_max)
}
\arguments{
\item{data_outcomes}{n x d matrix of discrete outcomes}

\item{s_max}{vector of length d, containing the maximum score for each test}
}
\value{
An n x d binary matrix (1 indicates observed, 0 is missing)
}
\description{
`get_R_matrix()` returns the n x d missing pattern matrix R
}
