% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gendata.R
\name{makeMiss}
\alias{makeMiss}
\title{Adds missingness to a generated complete case data set}
\usage{
makeMiss(dataset, eta = 0)
}
\arguments{
\item{dataset}{list of generated data set containing X, Y, Z}

\item{eta}{parameter that controls missingness, eta -> inf => less missingness,
eta -> 0 => most missingness}
}
\value{
A list of X (n x p) covariate matrix, Y (n x d) discrete outcome
matrix, Y_obs (n x d) discrete outcome matrix with NaNs,
Z (n x 1) vector of class labels
}
\description{
`makeMiss()` returns a generated data set with missing data
}
