% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_est.R
\name{multiple_impute}
\alias{multiple_impute}
\title{Multiple imputation
(Algorithm 2 from paper)}
\usage{
multiple_impute(
  data_cov,
  data_outcomes,
  theta,
  beta,
  s_max,
  M_imp = 50,
  stacked = TRUE
)
}
\arguments{
\item{data_cov}{n x p matrix of covariates}

\item{data_outcomes}{n x d matrix of discrete outcomes}

\item{theta}{K x d matrix of binomial theta params}

\item{beta}{(K-1) x (p+1) matrix of beta (weight) initial params}

\item{stacked}{True/False flag on whether to stack the imputed data sets}

\item{M_imps}{number of multiple imputations}
}
\value{
All M imputed data sets
}
\description{
`multiple_impute()` returns a list containing the beta_star,
theta_star, and log_likelihood evaluated at the convergence point
}
