% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{predict_probs}
\alias{predict_probs}
\title{Predict class probabilities w_k(x) for each obs}
\usage{
predict_probs(beta_matrix, cov_matrix, log = FALSE)
}
\arguments{
\item{beta_matrix}{(K-1) x (p+1) matrix of beta (weight) parameters}

\item{cov_matrix}{n x p matrix of covariates for all obs}

\item{log}{True/False flag to determine whether to return probs in log form}
}
\value{
An n x K matrix of (log) probabilities for each class and obs
}
\description{
`predict_probs()` retrieves w_k(x) for each obs
}
